+++ txcache/transactionsHeapItem.go
@@ -1,0 +1,7 @@
+package txcache
+
+type TransactionsHeapItem struct {
+	senderIndex      int
+	transactionIndex int
+	transaction      *WrappedTransaction
+}
